Use precompiled headers for libopenrct2 with MSVC

+++ src/openrct2/drawing/LightFX.cpp
@@ -384,11 +384,11 @@ void lightfx_prepare_light_list()
                 continue;
             }
 
-            entry->LightIntensity = std::min<uint32_t>(
-                0xFF, (entry->LightIntensity * lightIntensityOccluded) / (totalSamplePoints * 100));
+            entry->LightIntensity = static_cast<uint8_t>(
+                std::min<uint32_t>(0xFF, (entry->LightIntensity * lightIntensityOccluded) / (totalSamplePoints * 100)));
         }
-        entry->LightIntensity = std::max<uint32_t>(
-            0x00, entry->LightIntensity - static_cast<int8_t>(_current_view_zoom_front) * 5);
+        entry->LightIntensity = static_cast<uint8_t>(
+            std::max<uint32_t>(0x00, entry->LightIntensity - static_cast<int8_t>(_current_view_zoom_front) * 5));
 
         if (_current_view_zoom_front > ZoomLevel{ 0 })
         {
@@ -1015,7 +1015,7 @@ static uint8_t mix_light(uint32_t a, uint32_t b, uint32_t intensity)
     intensity = intensity * 6;
     uint32_t bMul = (b * intensity) >> 8;
     uint32_t ab = a + bMul;
-    uint8_t result = std::min<uint32_t>(255, ab);
+    uint8_t result = static_cast<uint8_t>(std::min<uint32_t>(255, ab));
     return result;
 }
